+++ Source/nz/net/ultraq/thymeleaf/LayoutDialect.java
@@ -20,7 +20,6 @@ import nz.net.ultraq.thymeleaf.decorator.DecoratorProcessor;
 import nz.net.ultraq.thymeleaf.decorator.TitlePatternProcessor;
 import nz.net.ultraq.thymeleaf.include.IncludeProcessor;
 import nz.net.ultraq.thymeleaf.include.ReplaceProcessor;
-import nz.net.ultraq.thymeleaf.include.SubstituteByProcessor;
 
 import org.thymeleaf.dialect.AbstractDialect;
 import org.thymeleaf.processor.IProcessor;
@@ -59,7 +58,6 @@ public class LayoutDialect extends AbstractDialect {
 		processors.add(new DecoratorProcessor());
 		processors.add(new IncludeProcessor());
 		processors.add(new ReplaceProcessor());
-		processors.add(new SubstituteByProcessor());
 		processors.add(new FragmentProcessor());
 		processors.add(new TitlePatternProcessor());
 		return processors;